feat(task-102): added aop @AdminOnly to provide access to admin commands to subscribers wih role admin/ added actuator and custom metric/ added open feign connection to OpenMeteo api and feature to notify subscribers about weather forecast

+++ src/main/java/com/sport/service/services/SubscriberService.java
@@ -19,7 +19,7 @@ public interface SubscriberService {
 
     long getUsersCount();
 
-    int getSubscriptionsCount();
+    long getSubscriptionsCount();
 
     List<Subscriber> findAll();
 }
